chore: refactor models helper methods
