+++ usecases/modules/modules.go
@@ -18,14 +18,13 @@ import (
 	"slices"
 	"sync"
 
-	"github.com/weaviate/weaviate/entities/dto"
-	schemachecks "github.com/weaviate/weaviate/entities/schema/checks"
-
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"github.com/tailor-inc/graphql"
 	"github.com/tailor-inc/graphql/language/ast"
+	"github.com/weaviate/weaviate/entities/dto"
 	"github.com/weaviate/weaviate/entities/models"
+	"github.com/weaviate/weaviate/entities/modelsext"
 	"github.com/weaviate/weaviate/entities/modulecapabilities"
 	"github.com/weaviate/weaviate/entities/moduletools"
 	"github.com/weaviate/weaviate/entities/schema"
@@ -927,7 +926,7 @@ func (p *Provider) getTargetVector(class *models.Class, params interface{}) ([]s
 		return nearParam.GetTargetVectors(), nil
 	}
 	if class != nil {
-		if schemachecks.HasLegacyVectorIndex(class) {
+		if modelsext.ClassHasLegacyVectorIndex(class) {
 			return []string{""}, nil
 		}
