+++ usecases/objects/auto_schema.go
@@ -21,6 +21,7 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/sirupsen/logrus"
+	"github.com/weaviate/weaviate/entities/modelsext"
 	"github.com/weaviate/weaviate/entities/vectorindex"
 
 	"github.com/weaviate/weaviate/entities/additional"
@@ -143,7 +144,7 @@ func (m *autoSchemaManager) createClass(ctx context.Context, principal *models.P
 		Properties:  properties,
 		Description: "This property was generated by Weaviate's auto-schema feature on " + now.Format(time.ANSIC),
 		VectorConfig: map[string]models.VectorConfig{
-			schema.DefaultNamedVectorName: {
+			modelsext.DefaultNamedVectorName: {
 				VectorIndexType: vectorindex.DefaultVectorIndexType,
 				Vectorizer:      map[string]any{m.defaultVectorizer: map[string]any{}},
 			},
