+++ usecases/objects/validation/vector_validation.go
@@ -16,8 +16,7 @@ import (
 	"fmt"
 
 	"github.com/weaviate/weaviate/entities/models"
-	"github.com/weaviate/weaviate/entities/schema"
-	schemachecks "github.com/weaviate/weaviate/entities/schema/checks"
+	"github.com/weaviate/weaviate/entities/modelsext"
 )
 
 func (v *Validator) vector(ctx context.Context, class *models.Class,
@@ -25,16 +24,16 @@ func (v *Validator) vector(ctx context.Context, class *models.Class,
 ) error {
 	// In case the schema has a legacy vector index, we need to check if there is a default named vector which
 	// is used to transport legacy vector via the named vector API.
-	if defaultVec, ok := incomingObject.Vectors[schema.DefaultNamedVectorName]; ok && schemachecks.HasLegacyVectorIndex(class) {
+	if defaultVec, ok := incomingObject.Vectors[modelsext.DefaultNamedVectorName]; ok && modelsext.ClassHasLegacyVectorIndex(class) {
 		vec, ok := defaultVec.([]float32)
 		if !ok {
-			return fmt.Errorf("vector %s has to be a float32 array", schema.DefaultNamedVectorName)
+			return fmt.Errorf("vector %s has to be a float32 array", modelsext.DefaultNamedVectorName)
 		}
 		incomingObject.Vector = vec
-		delete(incomingObject.Vectors, schema.DefaultNamedVectorName)
+		delete(incomingObject.Vectors, modelsext.DefaultNamedVectorName)
 	}
 
-	if !schemachecks.HasLegacyVectorIndex(class) && len(incomingObject.Vector) > 0 {
+	if !modelsext.ClassHasLegacyVectorIndex(class) && len(incomingObject.Vector) > 0 {
 		// if there is only one named vector we can assume that the single vector
 		if len(class.VectorConfig) == 1 {
 			namedVectorName := ""
