+++ usecases/schema/class.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
-	schemachecks "github.com/weaviate/weaviate/entities/schema/checks"
+	"github.com/weaviate/weaviate/entities/modelsext"
 
 	"github.com/weaviate/weaviate/adapters/repos/db/inverted/stopwords"
 	"github.com/weaviate/weaviate/entities/backup"
@@ -331,7 +331,7 @@ func (h *Handler) setClassDefaults(class *models.Class, globalCfg replication.Gl
 	// set legacy vector index defaults only when:
 	// 	- no target vectors are configured
 	//  - OR, there are target vectors configured AND there is a legacy vector configured
-	if !hasTargetVectors(class) || schemachecks.HasLegacyVectorIndex(class) {
+	if !hasTargetVectors(class) || modelsext.ClassHasLegacyVectorIndex(class) {
 		if class.Vectorizer == "" {
 			class.Vectorizer = h.config.DefaultVectorizerModule
 		}
@@ -639,7 +639,7 @@ func setInvertedConfigDefaults(class *models.Class) {
 func (h *Handler) validateCanAddClass(ctx context.Context, class *models.Class, classGetterWithAuth func(string) (*models.Class, error),
 	relaxCrossRefValidation bool,
 ) error {
-	if schemachecks.HasLegacyVectorIndex(class) && len(class.VectorConfig) > 0 {
+	if modelsext.ClassHasLegacyVectorIndex(class) && len(class.VectorConfig) > 0 {
 		return fmt.Errorf("creating a class with both a class level vector index and named vectors is forbidden")
 	}
 
@@ -776,7 +776,7 @@ func (h *Handler) validatePropertyIndexing(prop *models.Property) error {
 }
 
 func (h *Handler) validateVectorSettings(class *models.Class) error {
-	if schemachecks.HasLegacyVectorIndex(class) {
+	if modelsext.ClassHasLegacyVectorIndex(class) {
 		if err := h.validateVectorIndexType(class.VectorIndexType); err != nil {
 			return err
 		}
